… server/uptime-kuma-server.js
… src/mixins/socket.js
… src/pages/EditMonitor.vue
… src/util.js
… src/util.ts