@CommanderStorm @louislam 
1) fix loop to async
2) query n+1 problem

Tick the checkbox if you understand [x]:
- [x] I have read and understand the pull request rules.

# Description

Resolves https://github.com/louislam/uptime-kuma/issues/5129

testing results:
monitor url count: 415
current: 
cpu: 1  memory 2 gb: 1 min 44 sec

after changes

database: aws rds marinadb db.t3.micro(memory: 1gb, cpu: 2vcpu)

with docker run on mac:
cpu: 0.5  memory 1 gb: 21 sec
cpu:   1  memory 2 gb: 11 sec
cpu: 1.5  memory 3 gb: 10 sec
cpu:   2  memory 4 gb: 10 sec

with ESC Farget(Linux/ARM64):
cpu: 0.5  memory 1 gb: 31 sec
cpu: 1  memory 2 gb: 6 sec
cpu: 2  memory 4 gb: 5 sec
Fixes #3494 

## Type of change

- Performance improvement

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I ran ESLint and other linters for modified files
- [x] I have performed a self-review of my own code and tested it
- [x] I have commented my code, particularly in hard-to-understand areas (including JSDoc for methods)
- [x] My changes generates no new warnings
- [ ] My code needed automated testing. I have added them (this is optional task)

## Screenshots (if any)
<img width="1036" alt="1cpu-2gbram" src="https://github.com/user-attachments/assets/99796f1b-ba2f-4ea6-b7d5-c77e4c406c8b">