@vishalsabhaya 2024-08-19
@CommanderStorm @louislam 

I debug & fix performance issues for insert/update/delete/pause/resume monitor.

current:
   page load time + 500ms to 1 sec
   it always fetches the latest data from the database & builds json. i felt that it was extra time.
   
  after changes:
   < 700ms
   
   I just updated the target record with the front end instead of all records.