@mabed-fr 2024-08-21
Can you give me the docker command to test your PR? I encounter the problem with 1000 monitor (4vcpu / 16go RAM)