@vishalsabhaya 2024-08-22
@CommanderStorm 

Thank you for your effert to review my pr. i am really appreciate your hard work for community.
I can't understand below comment can you elaborate?

> While the ~16x(!) performance improvement in this PR is great, I don't quite like how you implemented some parts.
> (most parts are fine)
> 
> preloadData is quite clunky
> => please move this into a single object per toJSON that we pass to the function instead of the current way of passing a lookup-object into the function.
> The current way is quite likely error prown and without tests whatsoever way in this part extremely risky.
> 
> Only passing an object also has the advantage of allowing better type-checking via adding what is passed in to the doc-comment.
>