@vishalsabhaya 2024-08-22
@CommanderStorm  @louislam 

I found 1 more thing. it will improve more seconds.

marked realtime dot is maximum 43. why we are fetching 100 records for each monitor? is there any reason?
currently I am testing with 834 monitor
limit 100 takes: 13 sec
limit 43 takes: 7 sec

<img width="763" alt="image" src="https://github.com/user-attachments/assets/9189d9fc-77cc-47df-9480-2edba88bffd2">

![image](https://github.com/user-attachments/assets/9f4934a9-fffa-4441-b82e-f3296a7effee)