@CommanderStorm 2024-08-23
>  is there any reason?

I can imagine two:

- What happens if you zoom out?
- What happens on high-dpi devices?

![image](https://github.com/user-attachments/assets/03d42549-fe20-42f5-ae6a-d25c15493018)

If we change that part, lets please do this in a second PR ^^