@vishalsabhaya 2024-08-26
> > is there any reason?
> 
> I can imagine two:
> 
> * What happens if you zoom out?
> * What happens on high-dpi devices?
> 
> ![image](https://private-user-images.githubusercontent.com/26258709/360926613-03d42549-fe20-42f5-ae6a-d25c15493018.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjQ2MzUwNTQsIm5iZiI6MTcyNDYzNDc1NCwicGF0aCI6Ii8yNjI1ODcwOS8zNjA5MjY2MTMtMDNkNDI1NDktZmUyMC00MmY1LWFlNmEtZDI1YzE1NDkzMDE4LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA4MjYlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwODI2VDAxMTIzNFomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTIyODE0NmM1ZjQ2NTFlYjIxN2YyNmNlZTUxYzVjYmI1NjI4NjVhYzJhNTY5MDc2NGZiNTNiYTc5Yzg3YTU0ZjAmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JmFjdG9yX2lkPTAma2V5X2lkPTAmcmVwb19pZD0wIn0.M-oY9rPLStfK8R31evcSeiYnx03hydH-svPsrHxeXtI)
> 
> If we change that part, lets please do this in a second PR ^^

yes, you are correct. i didn't think about it. we will do in next pr. but we should restrict max span. because for detail information already provided graph.