@vishalsabhaya 2024-09-04
@CommanderStorm 
i seen something strange. did you notice ever?

i did testing with 800 records.

when I remove below line. all monitoring is working fine but current case most of monitor is stopped. monitoring data is not updating so monitor is not starting.
await sleep(getRandomInt(300, 1000));
![image](https://github.com/user-attachments/assets/36aef699-cb34-4bcf-ae4a-3276c65384bb)


startMonitors does not have await that's why?
![image](https://github.com/user-attachments/assets/4667cd4b-07b0-4417-b951-5704e7c988a0)