@CommanderStorm 2024-09-04
Sorry, above comment does not quite parse for me.

As far as I understand that line, this just means that the monitors are started one after another with random intervals in between.

The way this is executed, starting monitors should happen after the server started listening, but must not have completely/partially completed before starting the `versionCheck`.
That this is not done in a Promise.all makes a [thundering heard problem](https://en.m.wikipedia.org/wiki/Thundering_herd_problem) less likely. The speed of starting them is constant though.

As far as I understand async js (feel free to correct), that sounds partially reasonable.
Changing the monitor startup to be smarter can be done, but let's separate this Into its own PR. ^^