@CommanderStorm 2024-09-30
Sorry, got back from vacation on monday and am still picking up the pieces work wise and had not time for such a large PR.
I can answer issues or merge small PRs on the train but for this one, I need an hour to look if everything works as intended.